redraw message list if added message not in last index

+++ Models/MessageEntry.cs
@@ -7,6 +7,6 @@ public sealed class MessageEntry
     public string Name { get; set; }
     public Color NameColor { get; set; }
     public string Text { get; set; }
-    public uint TimeStamp { get; set; }
+    public ulong TimeStamp { get; set; }
     public string Time { get; set; }
 }
